GetFoto
Get Foto and refactor api request
