+++ src/components/ProfileAccount.js
@@ -3,59 +3,19 @@ import { StyleSheet, Text, View } from 'react-native';
 import { Avatar, BottomSheet, ListItem } from 'react-native-elements';
 import { widthPercentageToDP as wp, heightPercentageToDP as hp } from 'react-native-responsive-screen';
 import { launchCameraAsync, MediaTypeOptions } from 'expo-image-picker';
-
-import ApiConfig from '../api/ApiConfig';
-import { Api, username, password } from '../api/UrlApi';
 import { loadImageFromGallery, askForPermission } from '../helpers/Permissions';
 import { ButtonCamera, ButtonGallery, ButtonCancel } from './BottomSheet';
+import { apiPostFoto, apiGetFoto } from '../api/ApiRequest';
 
 const ProfileAccount = () => {
   const [isVisible, setIsVisible] = useState(false);
   const [photografy, setphotografy] = useState(null);
-
-  const apiRequest = uri => {
-    let numero = '19190780';
-    let uriImage = uri;
-    ApiConfig.post('auth/local', {
-      identifier: username,
-      password: password,
-    })
-      .then(res => {
-        return res.data;
-      })
-      .then(user => {
-        const formData = new FormData();
-        formData.append(
-          'data',
-          JSON.stringify({
-            ncontrol: numero,
-          }),
-        );
-        let filename = numero + '-image.jpg';
-        formData.append('files.Foto', {
-          uri: uriImage,
-          name: filename,
-          type: 'image/jpeg',
-        });
-
-        if (user.jwt) {
-          const response = fetch(`${Api}/credenciales`, {
-            method: 'POST',
-            body: formData,
-            headers: {
-              Authorization: `Bearer ${user.jwt}`,
-            },
-          });
-          const data = response.json();
-          console.log('data', data);
-        }
-      });
-  };
-
+  let ncontrol = '18190687';
+  const url = apiGetFoto(ncontrol);
   const openImage = async () => {
     const result = await loadImageFromGallery([1, 1]);
     setphotografy(result);
-    apiRequest(result.image);
+    apiPostFoto(result.image, ncontrol);
   };
 
   const takeImage = async () => {
@@ -73,7 +33,7 @@ const ProfileAccount = () => {
 
       setphotografy({ localUri: image.uri });
       if (!image.cancelled) {
-        apiRequest(image.uri);
+        apiPostFoto(image.uri, ncontrol);
       }
     }
   };
@@ -105,7 +65,7 @@ const ProfileAccount = () => {
   return (
     <View style={[styles.containerContent, styles.mb20]}>
       <View style={[styles.photoProfile, styles.mb20]}>
-        <Avatar onPress={() => setIsVisible(true)} size={160} source={require('../assets/images/user.png')} />
+        <Avatar onPress={() => setIsVisible(true)} size={160} source={url} />
         <BottomSheet modalProps={{}} isVisible={isVisible}>
           {list.map((l, i) => (
             <ListItem key={i} containerStyle={l.containerStyle} onPress={l.onPress}>
